reviews organized in their columns and components refactored to have own directory with styles
