+++ client/components/feedbackCollector/feedbackCollectorComponent.styles.jsx
@@ -1,0 +1,7 @@
+export const FeedBackOrganizer = styled.div`
+  display: flex;
+`;
+
+export const FeedBackOrganizerChild = styled.div`
+display: flex;
+`;
